chore(): add semantic routing
