+++ contracts/deploy/deploy.ts
@@ -1,11 +1,15 @@
 import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
-import { parseUnits, Wallet } from "ethers";
+import { parseUnits } from "ethers";
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deployments, getNamedAccounts, network, ethers } = hre;
   const { deploy } = deployments;
   const { deployer } = await getNamedAccounts();
+
+  // 🛑 Helper: Wait for N milliseconds
+  const delay = (ms: number) => new Promise((res) => setTimeout(res, ms));
+  const RPC_DELAY = 5000; // 5 second pause between actions
 
   console.log(`\n🚀 Deploying Slice to network: ${network.name}`);
 
@@ -16,6 +20,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   };
 
   let usdcAddress = MOCK_USDC_ADDRESSES[network.name];
+
   if (network.name === "hardhat" || network.name === "localhost") {
     const mock = await deploy("MockUSDC", { from: deployer, args: [], log: true });
     usdcAddress = mock.address;
@@ -23,27 +28,28 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
 
   if (!usdcAddress) throw new Error("Missing USDC Address");
 
-  // --- 2. MINT FUNDS (SMART CHECK) ---
+  // --- 2. CHECK & MINT FUNDS ---
   const signer = await ethers.getSigner(deployer);
   const usdc = await ethers.getContractAt("MockUSDC", usdcAddress, signer);
 
-  try {
-    const myBalance = await usdc.balanceOf(deployer);
-    const mintAmount = parseUnits("10000", 6);
+  if (network.name === "hardhat" || network.name === "localhost" || network.name === "baseSepolia") {
+    try {
+      const myBalance = await usdc.balanceOf(deployer);
+      const mintAmount = parseUnits("10000", 6);
 
-    if (myBalance < mintAmount) {
-      console.log("   🔄 Balance low. Minting 10,000 USDC...");
-      try {
-        await (await (usdc as any).mint(deployer, mintAmount)).wait();
-      } catch {
-        await (await (usdc as any).mint(mintAmount)).wait();
+      if (myBalance < mintAmount) {
+        console.log("   🔄 Balance low. Minting 10,000 USDC...");
+        try {
+          await (await (usdc as any).mint(deployer, mintAmount)).wait();
+        } catch {
+          await (await (usdc as any).mint(mintAmount)).wait();
+        }
+        console.log("   ✅ Mint Successful");
+        await delay(RPC_DELAY); // Pause after mint
       }
-      console.log("   ✅ Mint Successful");
-    } else {
-      console.log("   Info: Deployer already has sufficient USDC. Skipping mint.");
+    } catch (e) {
+      console.log("   ⚠️  Mint check failed/skipped");
     }
-  } catch (e) {
-    console.log("   ⚠️  Mint check failed/skipped");
   }
 
   // --- 3. DEPLOY SLICE ---
@@ -55,36 +61,50 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   });
   console.log(`✅ Slice deployed at: ${sliceDeploy.address}`);
 
+  await delay(RPC_DELAY); // Pause after deploy
+
   const slice = await ethers.getContractAt("Slice", sliceDeploy.address, signer);
 
   // --- 4. SEED DISPUTES ---
-  if (process.env.SEED_DISPUTES === "true" || network.name === "hardhat" || network.name === "baseSepolia") {
+  const shouldSeed =
+    process.env.SEED_DISPUTES === "true" || network.name === "hardhat" || network.name === "baseSepolia";
+
+  if (shouldSeed) {
     console.log("\n🌱 Seeding initial disputes...");
 
-    // 1. Get Deployer Nonce
-    let currentNonce = await signer.getNonce("pending");
-    console.log(`   🔄 Starting Deployer Nonce: ${currentNonce}`);
+    // Wait for indexing
+    console.log("   ⏳ Waiting 10s for network indexing...");
+    await delay(10000);
 
-    // 2. Setup Defender & Defender Nonce
-    const defenderWallet = Wallet.createRandom().connect(ethers.provider);
-    let defenderNonce = 0; // Starts at 0 for a fresh wallet
+    const signers = await ethers.getSigners();
+    const defenderWallet = signers[1];
+
+    if (!defenderWallet) {
+      throw new Error("Defender account not found.");
+    }
     console.log(`   🤖 Defender: ${defenderWallet.address}`);
 
-    // 3. Fund Defender (ETH)
-    console.log("   💸 Sending ETH to defender...");
-    await (
-      await signer.sendTransaction({
-        to: defenderWallet.address,
-        value: parseUnits("0.000005", 18),
-        nonce: currentNonce++,
-      })
-    ).wait();
+    // Fund Defender (ETH)
+    const defenderBalance = await ethers.provider.getBalance(defenderWallet.address);
+    if (defenderBalance < parseUnits("0.00001", 18)) {
+      console.log("   💸 Sending ETH to defender...");
+      await (
+        await signer.sendTransaction({
+          to: defenderWallet.address,
+          value: parseUnits("0.00005", 18),
+        })
+      ).wait();
+      await delay(RPC_DELAY); // Pause after funding ETH
+    }
 
-    // 4. Fund Defender (USDC)
-    console.log("   💸 Sending USDC to defender...");
-    await (await usdc.transfer(defenderWallet.address, parseUnits("500", 6), { nonce: currentNonce++ })).wait();
+    // Fund Defender (USDC)
+    const defenderUSDC = await usdc.balanceOf(defenderWallet.address);
+    if (defenderUSDC < parseUnits("100", 6)) {
+      console.log("   💸 Sending USDC to defender...");
+      await (await usdc.transfer(defenderWallet.address, parseUnits("500", 6))).wait();
+      await delay(RPC_DELAY); // Pause after funding USDC
+    }
 
-    // Data
     const ROOT_CID = "bafybeifa6gsnklvyvepp45ilf4ngc5o3ndydq7zxcdgrfybxs6flts6mdi";
     const disputes = [
       { title: "Freelance", category: "Freelance", ipfsHash: `${ROOT_CID}/freelance.json` },
@@ -95,29 +115,35 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     const STAKE_AMOUNT = 1000000n; // 1 USDC
     const ONE_WEEK = 604800;
 
+    const sliceDef = slice.connect(defenderWallet);
+    const usdcDef = usdc.connect(defenderWallet);
+
     for (const d of disputes) {
       console.log(`   Detailed Setup for: "${d.title}"`);
 
-      // A. Approve (Claimer)
-      await (await usdc.approve(slice.target, STAKE_AMOUNT, { nonce: currentNonce++ })).wait();
+      // ---------------------------------------------------------
+      // A. CLAIMER ACTIONS
+      // ---------------------------------------------------------
 
-      // B. Create Dispute
-      const tx = await slice.createDispute(
-        {
-          claimer: deployer,
-          defender: defenderWallet.address,
-          category: d.category,
-          ipfsHash: d.ipfsHash,
-          jurorsRequired: 1,
-          paySeconds: ONE_WEEK,
-          evidenceSeconds: ONE_WEEK,
-          commitSeconds: ONE_WEEK,
-          revealSeconds: ONE_WEEK,
-        },
-        { nonce: currentNonce++ },
-      );
+      console.log("     ... Claimer Approving");
+      const approveTx = await usdc.approve(slice.target, STAKE_AMOUNT);
+      await approveTx.wait();
+      await delay(RPC_DELAY); // <--- DELAY ADDED
 
-      const receipt = await tx.wait();
+      console.log("     ... Claimer Creating Dispute");
+      const createTx = await slice.createDispute({
+        claimer: deployer,
+        defender: defenderWallet.address,
+        category: d.category,
+        ipfsHash: d.ipfsHash,
+        jurorsRequired: 1n,
+        paySeconds: ONE_WEEK,
+        evidenceSeconds: ONE_WEEK,
+        commitSeconds: ONE_WEEK,
+        revealSeconds: ONE_WEEK,
+      });
+      const receipt = await createTx.wait();
+      await delay(RPC_DELAY); // <--- DELAY ADDED
 
       // Find ID
       let disputeId = 0;
@@ -137,42 +163,40 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
         console.error("❌ Could not find DisputeCreated event. Skipping payment.");
         continue;
       }
-      console.log(`      ✅ Created Dispute #${disputeId}`);
+      console.log(`       ✅ Created Dispute #${disputeId}`);
 
-      // C. Pay (Claimer)
-      await (await slice.payDispute(disputeId, { nonce: currentNonce++ })).wait();
-      console.log(`      💰 Claimer Paid`);
+      console.log("     ... Claimer Paying");
+      const payClaimerTx = await slice.payDispute(disputeId);
+      await payClaimerTx.wait();
+      await delay(RPC_DELAY); // <--- DELAY ADDED
+      console.log(`       💰 Claimer Paid`);
 
-      // D. Pay (Defender) - WITH MANUAL DEFENDER NONCE
-      const sliceDef = slice.connect(defenderWallet);
-      const usdcDef = usdc.connect(defenderWallet);
+      // ---------------------------------------------------------
+      // B. DEFENDER ACTIONS
+      // ---------------------------------------------------------
 
-      // D1. Approve
-      await (
-        await usdcDef.approve(
-          slice.target,
-          STAKE_AMOUNT,
-          { nonce: defenderNonce++ }, // <--- Manual Defender Nonce
-        )
-      ).wait();
+      console.log(`     ... Defender Approving`);
+      const defApproveTx = await usdcDef.approve(slice.target, STAKE_AMOUNT);
+      await defApproveTx.wait();
+      await delay(RPC_DELAY); // <--- DELAY ADDED
 
-      // D2. Pay
-      await (
-        await sliceDef.payDispute(
-          disputeId,
-          { nonce: defenderNonce++ }, // <--- Manual Defender Nonce
-        )
-      ).wait();
+      console.log(`     ... Defender Paying`);
+      const defPayTx = await sliceDef.payDispute(disputeId);
+      await defPayTx.wait();
+      await delay(RPC_DELAY); // <--- DELAY ADDED
 
-      console.log(`      💰 Defender Paid -> Status: COMMIT`);
+      console.log(`       💰 Defender Paid -> Status: COMMIT`);
     }
   }
 
   // --- 5. VERIFY ---
   if (network.name !== "hardhat" && network.name !== "localhost" && process.env.BASESCAN_API_KEY) {
+    console.log("   Verifying contract...");
     try {
       await hre.run("verify:verify", { address: sliceDeploy.address, constructorArguments: [usdcAddress] });
-    } catch (e) {}
+    } catch (e) {
+      console.log("   Verification skipped or failed.");
+    }
   }
 };
