+++ src/app/juror/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function JurorPage() {
+  redirect("/juror/tasks");
+}
